Datasets/evalplus/Java/src/test/java/com/example/HumanEval_38Test.java
package com.example;
import static org.junit.Assert.assertEquals;
import org.junit.Test;

public class HumanEval_38Test {

    @Test
    public void testDecodeCyclic_0() {
        assertEquals("fuzzplzjftcllmdy", HumanEval_38.decodeCyclic("uzfplzjfzcltmdly"));
    }

    @Test
    public void testDecodeCyclic_1() {
        assertEquals("ynzaegrghwzqzrdkxcn", HumanEval_38.decodeCyclic("nzyegaghrzqwrdzxckn"));
    }

    // Add similar test methods for the other test cases...

    @Test
    public void testDecodeCyclic_102() {
        assertEquals("3126459780", HumanEval_38.decodeCyclic("1234567890"));
    }

}
// End of Test
